Day_1_LangChain Models/2.ChatModels/2_chatmodel_anthrophic.py
#CLOSED SOURCE

# claude ai by anthrophic company
#so import that library
from langchain_anthropic import ChatAnthropic
from dotenv import load_dotenv

load_dotenv()

model = ChatAnthropic(model="claude-opus-4-20250514")

result = model.invoke("What is the capital of India")

print(result)