Day_1_LangChain Models/3.EmbeddingModels/5_document_similarity_hf.py
from langchain_huggingface import HuggingFaceEmbeddings

from sklearn.metrics.pairwise import cosine_similarity
from dotenv import load_dotenv
import numpy as np

load_dotenv()

# Use Hugging Face embedding model (open source, local)
embedding = HuggingFaceEmbeddings(model_name='sentence-transformers/all-MiniLM-L6-v2')

documents = [
    "Virat Kohli is an Indian cricketer known for his aggressive batting and leadership.",
    "MS Dhoni is a former Indian captain famous for his calm demeanor and finishing skills.",
    "Sachin Tendulkar, also known as the 'God of Cricket', holds many batting records.",
    "Rohit Sharma is known for his elegant batting and record-breaking double centuries.",
    "Jasprit Bumrah is an Indian fast bowler known for his unorthodox action and yorkers."
]

query = "tell me about virat kohli"

# Get embeddings
doc_embeddings = embedding.embed_documents(documents)
query_embedding = embedding.embed_query(query)

# Cosine similarity
scores = cosine_similarity([query_embedding], doc_embeddings)[0]

# Get best match
index, score = sorted(enumerate(scores), key=lambda x: x[1])[-1]

print(f"Query: {query}")
print(f"Best Match: {documents[index]}")
print(f"Similarity Score: {score:.4f}")
